Apply patch from Neil Brown.
[PATCH] nfs coherence: add interface to refresh mapping when we lock
 the file.

NFS mapping are only validated against the server when the mmap call
is made, not when accesses happen, even if byte-range locking is used.
So for multi-host coherences, we need to unmap and re-map the file
whenever we take out a lock.

This patch doesn't check for errors in the 'mmap' call ... maybe it
should.

Signed-off-by: NeilBrown <neilb@suse.com>